refactor: move DatabaseProperties to its own model
