+++ frontend/src/types/database.ts
@@ -1,0 +1,11 @@
+export interface DatabaseProperties {
+  id: number;
+  name: string;
+  engine: string;
+  createdAt: Date;
+}
+
+export type UpstreamDatabaseProperties = Pick<
+  DatabaseProperties,
+  "name" | "engine"
+>;
